requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
numpy==2.1.2
xarray==2024.7.0
dask==2024.8.1
pandas==2.2.3
python-dotenv==1.0.1
requests==2.32.3
earthaccess==0.9.0
zarr==2.18.2
h5netcdf==1.3.0
httpx==0.27.0
aiofiles==24.1.0
pyarrow==21.0.0
